+++ changelog.md
@@ -3,7 +3,7 @@ All notable changes to Harmony will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Version 6
-#### [Unreleased]
+#### [6.0.0]
 - Expose `slot.gpt` to decouple GPT slot from Harmony slot.
 - Slot option `interstitial` changed to `outofpage`.
 - Slots are no longer defined in GPT until they are displayed. This